質問リストをコンポーネントに切り分け
