+++ src/features/questions/components/index.js
@@ -1,0 +1,3 @@
+import QuestionList from "./questionList";
+
+export { QuestionList };
